requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.3
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0
spacy>=3.7.2
nltk>=3.8.1
sentence-transformers>=2.3.1
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.3
pymongo>=4.6.1
motor>=3.3.2
sqlalchemy>=2.0.25
langchain>=0.1.4
langchain-community>=0.0.16
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
python-dotenv>=1.0.0
lime>=0.2.0.1
shap>=0.44.1
transformers>=4.37.2

streamlit>=1.30.0
plotly>=5.18.0
requests>=2.31.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
